requirements.txt

pyjwt==2.9.0
rdflib>=7.0.0
requests>=2.0.0
fastapi[standard]==0.112.0
uvicorn==0.30.6
pydantic==2.8.2
passlib[bcrypt]





